dynamic assignment of error based on payload
The new default behaviour of action creators is to set the error property
to: payload instanceof Error
Passing a boolean as the isError argument sets the error property
unconditionally for all created actions.
Passing a function as the isError argument gives you full control for
calculating the error property dynamically based on the full payload.
The factory can be customized to change the default isError function.